feat(shaders): added the gradual_shift_duration option to shaders
This commit adds the `gradual_shift_duration` option to shaders in
hyprshade.toml. Any shader with a `start_time` can use have this option.

The option will set scheduled steps to increment the `gradualPercentage`
glsl variable if the concerned shader is a templage.
If the shader is not a template or does not implement the
`gradualPercentage` variable, `gradual_shift_duration` will have no
effect.

The `blue-light-filter` shader is now a mustache template and the `gradualPercentage` variable is implemented for it.

The option was also added to the `test_config`